+++ src/components/ImageSlider.css
