+++ src/components/data/SliderData.js
