+++ src/containers/Clothing.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { SliderData } from "../components/data/SliderData";
+import ImageSlider from "../components/ImageSlider";
+import ProductModal from "../components/ProductModal";
+import ProductCard from "../components/ProductCard";
+
+function Clothing() {
+  return (
+    <div className="clothing">
+      <ImageSlider slides={SliderData} />
+    </div>
+  );
+}
+
+export default Clothing;
